add Cats category
I had to define an equivalence for functor. it turns out that defining functor
equivalence in the default way (having functors map the same morphism to the same
morphism) seems to require K to prove it to be an equivalence relation. I made
equivalence of functors defined up to isomorphism.