Add functors between (co)cone categories
This adds three functors between cone categories (and cocone categories,
respectively). All F₀ and F₁ fields are maps that already have been
defined in Categories.Diagram.Co(co)ne.Properties, and thus the Functor
definition is places there as well. Unfortunately, using the definition
of Cones (and Cocones) requires the import of
Categories.Category.Construction.Cones (and Cocones) in the
Cones.Properties module.

For `F : Functor J C`, the new functors are:

  - Every `G : Functor C D` lifts to a `Functor (Cones F) (Cones (G ∘F F))`
    and `Functor (Cocones F) (Cocones (G ∘F F))`
  - Every `G : Functor J′ J` lifts to a `Functor (Cones F) (Cones (F ∘F G))`
    and `Functor (Cocones F) (Cocones (F ∘F G))`
  - Every `G : Functor J C` and `α : NaturalTransformation F G` extends
    to a `Functor (Cones F) (Cones G)` and `Functor (Cocones G) (Cocones F)`

In Cocones.Properties, I've restricted the import of Cone.Properties to
the four maps that are actually used.